# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for Campwiz",
        "title": "Campwiz API",
        "contact": {
            "name": "Nokib Sarkar",
            "url": "https://github.com/nokibsarkar",
            "email": "nokibsarkar@gmail.com"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "1"
    },
    "host": "localhost:8081",
    "basePath": "/api/v2",
    "paths": {
        "/campaign/": {
            "get": {
                "description": "get all campaigns",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "List all campaigns",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "name": "ids",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Whether the campaign is closed (result have been finalized)",
                        "name": "isClosed",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Whether the campaign is hidden from the public list\nIf isHidden is true, then projectID is required",
                        "name": "isHidden",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "next",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "prev",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "This projectID is the project that campaigns belong to.then ProjectID is required\n\t\tIf the person is not an admin, then the project ID must match the project ID of the user",
                        "name": "projectId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseList-models_Campaign"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new campaign",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "Create a new campaign",
                "parameters": [
                    {
                        "description": "The campaign request",
                        "name": "campaignRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CampaignCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Campaign"
                        }
                    }
                }
            }
        },
        "/campaign/{campaignId}": {
            "get": {
                "description": "Get a single campaign",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "Get a single campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The campaign ID",
                        "name": "campaignId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "name": "includeProject",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "includeRoles",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "includeRoundRoles",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "includeRounds",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSingle-models_CampaignExtended"
                        }
                    }
                }
            }
        },
        "/campaign/{id}": {
            "post": {
                "description": "Update a campaign",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "Update a campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The campaign ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The campaign request",
                        "name": "campaignRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CampaignUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSingle-models_Campaign"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "consts.Language": {
            "type": "string",
            "enum": [
                "am",
                "ar",
                "as",
                "af",
                "bn",
                "zh",
                "da",
                "nl",
                "en",
                "fi",
                "fr",
                "de",
                "el",
                "gu",
                "ha",
                "hi",
                "ig",
                "it",
                "ja",
                "kn",
                "ko",
                "ml",
                "mr",
                "ne",
                "no",
                "or",
                "om",
                "fa",
                "pt",
                "pa",
                "ru",
                "sa",
                "st",
                "sn",
                "sd",
                "so",
                "es",
                "sw",
                "sv",
                "ta",
                "te",
                "ts",
                "tn",
                "tr",
                "ur",
                "-",
                "ve",
                "xh",
                "yo",
                "zu"
            ],
            "x-enum-varnames": [
                "Amharic",
                "Arabic",
                "Assamese",
                "Afrikaans",
                "Bangla",
                "Chinese",
                "Danish",
                "Dutch",
                "English",
                "Finnish",
                "French",
                "German",
                "Greek",
                "Gujarati",
                "Hausa",
                "Hindi",
                "Igbo",
                "Italian",
                "Japanese",
                "Kannada",
                "Korean",
                "Malayalam",
                "Marathi",
                "Nepali",
                "Norwegian",
                "Odia",
                "Oromo",
                "Persian",
                "Portuguese",
                "Punjabi",
                "Russian",
                "Sanskrit",
                "Sesotho",
                "Shona",
                "Sindhi",
                "Somali",
                "Spanish",
                "Swahili",
                "Swedish",
                "Tamil",
                "Telugu",
                "Tsonga",
                "Tswana",
                "Turkish",
                "Urdu",
                "Undefined",
                "Venda",
                "Xhosa",
                "Yoruba",
                "Zulu"
            ]
        },
...
        "models.Campaign": {
            "type": "object",
            "required": [
                "endDate",
                "language",
                "name",
                "startDate"
            ],
            "properties": {
                "campaignId": {
                    "description": "A unique identifier for the campaign, it should be custom defined",
                    "type": "string"
                },
                "createdAt": {
                    "description": "The time the campaign was created, it would be set automatically",
                    "type": "string"
                },
                "createdById": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "isPublic": {
                    "description": "Whether the campaign is shown in the public list",
                    "type": "boolean"
                },
                "language": {
                    "$ref": "#/definitions/consts.Language"
                },
                "latestRoundId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "project": {
                    "$ref": "#/definitions/models.Project"
                },
                "projectId": {
                    "type": "string"
                },
                "rounds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Round"
                    }
                },
                "rules": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.RoundStatus"
                }
            }
        },
        "models.CampaignExtended": {
            "type": "object",
            "required": [
                "endDate",
                "language",
                "name",
                "startDate"
            ],
            "properties": {
                "campaignId": {
                    "description": "A unique identifier for the campaign, it should be custom defined",
                    "type": "string"
                },
                "coordinators": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "createdAt": {
                    "description": "The time the campaign was created, it would be set automatically",
                    "type": "string"
                },
                "createdById": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "isPublic": {
                    "description": "Whether the campaign is shown in the public list",
                    "type": "boolean"
                },
                "language": {
                    "$ref": "#/definitions/consts.Language"
                },
                "latestRoundId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "project": {
                    "$ref": "#/definitions/models.Project"
                },
                "projectId": {
                    "type": "string"
                },
                "rounds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Round"
                    }
                },
                "rules": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.RoundStatus"
                }
            }
        },
...
        "models.MediaType": {
            "type": "string",
            "enum": [
                "ARTICLE",
                "BITMAP",
                "AUDIO",
                "VIDEO",
                "PDF"
            ],
            "x-enum-varnames": [
                "MediaTypeArticle",
                "MediaTypeImage",
                "MediaTypeAudio",
                "MediaTypeVideo",
                "MediaTypePDF"
            ]
        },
        "models.Project": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdById": {
                    "type": "string"
                },
                "logoUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "projectId": {
                    "type": "string"
                },
                "url": {
                    "description": "The URL of the project's website",
                    "type": "string"
                }
            }
        },
...
        "models.ResponseList-models_Campaign": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Campaign"
                    }
                },
                "next": {
                    "type": "string"
                },
                "prev": {
                    "type": "string"
                }
            }
        },
...
        "models.ResponseSingle-models_Campaign": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Campaign"
                }
            }
        },
        "models.ResponseSingle-models_CampaignExtended": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.CampaignExtended"
                }
            }
        },
...
        "models.Round": {
            "type": "object",
            "properties": {
                "allowJuryToParticipate": {
                    "type": "boolean"
                },
                "allowMultipleJudgement": {
                    "type": "boolean"
                },
                "allowedMediaTypes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MediaType"
                    }
                },
                "articleAllowCreations": {
                    "type": "boolean"
                },
                "articleAllowExpansions": {
                    "type": "boolean"
                },
                "articleMaximumSubmissionOfSameArticle": {
                    "type": "integer"
                },
                "articleMinimumAddedBytes": {
                    "type": "integer"
                },
                "articleMinimumAddedWords": {
                    "type": "integer"
                },
                "articleMinimumTotalBytes": {
                    "type": "integer"
                },
                "articleMinimumTotalWords": {
                    "type": "integer"
                },
                "audioMinimumDurationMilliseconds": {
                    "type": "integer"
                },
                "audioMinimumSizeBytes": {
                    "type": "integer"
                },
                "blacklist": {
                    "type": "string"
                },
                "campaignId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdById": {
                    "type": "string"
                },
                "dependsOnRoundId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "imageMinimumResolution": {
                    "type": "integer"
                },
                "imageMinimumSizeBytes": {
                    "type": "integer"
                },
                "isOpen": {
                    "type": "boolean"
                },
                "isPublicJury": {
                    "type": "boolean"
                },
                "jury": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "latestTaskId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "projectId": {
                    "type": "string"
                },
                "quorum": {
                    "type": "integer"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Role"
                    }
                },
                "roundId": {
                    "type": "string"
                },
                "secretBallot": {
                    "type": "boolean"
                },
                "serial": {
                    "type": "integer"
                },
                "startDate": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.RoundStatus"
                },
                "totalAssignments": {
                    "type": "integer"
                },
                "totalEvaluatedAssignments": {
                    "type": "integer"
                },
                "totalEvaluatedSubmissions": {
                    "type": "integer"
                },
                "totalSubmissions": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/models.EvaluationType"
                },
                "videoMinimumDurationMilliseconds": {
                    "type": "integer"
                },
                "videoMinimumResolution": {
                    "type": "integer"
                },
                "videoMinimumSizeBytes": {
                    "type": "integer"
                }
            }
        },
        "models.RoundStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "IMPORTING",
                "DISTRIBUTING",
                "EVALUATING",
                "REJECTED",
                "CANCELLED",
                "PAUSED",
                "SCHEDULED",
                "ACTIVE",
                "COMPLETED"
            ],
            "x-enum-varnames": [
                "RoundStatusPending",
                "RoundStatusImporting",
                "RoundStatusDistributing",
                "RoundStatusEvaluating",
                "RoundStatusRejected",
                "RoundStatusCancelled",
                "RoundStatusPaused",
                "RoundStatusScheduled",
                "RoundStatusActive",
                "RoundStatusCompleted"
            ]
        },
...
        "services.CampaignCreateRequest": {
            "type": "object",
            "required": [
                "endDate",
                "language",
                "name",
                "startDate"
            ],
            "properties": {
                "coordinators": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "isPublic": {
                    "description": "Whether the campaign is shown in the public list",
                    "type": "boolean"
                },
                "language": {
                    "$ref": "#/definitions/consts.Language"
                },
                "name": {
                    "type": "string"
                },
                "organizers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "projectId": {
                    "type": "string"
                },
                "rules": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.RoundStatus"
                }
            }
        },
        "services.CampaignUpdateRequest": {
            "type": "object",
            "required": [
                "endDate",
                "language",
                "name",
                "startDate"
            ],
            "properties": {
                "coordinators": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "isPublic": {
                    "description": "Whether the campaign is shown in the public list",
                    "type": "boolean"
                },
                "language": {
                    "$ref": "#/definitions/consts.Language"
                },
                "name": {
                    "type": "string"
                },
                "organizers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "projectId": {
                    "type": "string"
                },
                "rules": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.RoundStatus"
                }
            }
        },
...
    }
}